+++ cli/running/running_test.go
@@ -20,6 +20,7 @@ func Test_CollectInstances(t *testing.T) {
 	require.NoError(t, err)
 	require.Equal(t, 1, len(instances))
 	require.Equal(t, InstanceCtx{
+		AppDir:         "testdata/instances_enabled/script",
 		AppName:        "script",
 		InstName:       "script",
 		InstanceScript: "testdata/instances_enabled/script.lua",
@@ -83,7 +84,7 @@ func Test_CollectInstances(t *testing.T) {
 	require.NoError(t, os.Chmod(instancesEnabledPath, 0666))
 	instances, err = CollectInstances("script", instancesEnabledPath)
 	assert.ErrorContains(t, err, "script.lua: permission denied")
-	assert.Equal(t, 0, len(instances))
+	assert.Equal(t, 1, len(instances))
 	require.NoError(t, os.Chmod(instancesEnabledPath, 0755))
 }
 
@@ -139,16 +140,17 @@ func Test_collectAppDirFiles(t *testing.T) {
 }
 
 func Test_collectInstancesForApps(t *testing.T) {
-	clusterAppName := "cluster_app"
+	appName := "cluster_app"
+	instancesEnabled, err := filepath.Abs("./testdata/instances_enabled")
+	require.NoError(t, err)
+	appLocation := filepath.Join(instancesEnabled, appName)
 	apps := []util.AppListEntry{
 		{
-			Name:     "cluster_app",
-			Location: "./testdata/instances_enabled/cluster_app",
+			Name: appName,
 		},
 	}
-	cfgDir := "/etc/tarantool/"
 	cliOpts := configure.GetDefaultCliOpts()
-	cliOpts.Env.InstancesEnabled = "./testdata/instances_enabled/"
+	cliOpts.Env.InstancesEnabled = instancesEnabled
 	instances, err := collectInstancesForApps(apps, cliOpts, "/etc/tarantool/")
 	require.NoError(t, err)
 	assert.Equal(t, 3, len(instances))
@@ -157,46 +159,49 @@ func Test_collectInstancesForApps(t *testing.T) {
 	for _, inst := range instances {
 		switch inst.InstName {
 		case "instance-001":
-			assert.Equal(t, filepath.Join(cfgDir, "var", "lib", clusterAppName, "instance-001"),
+			assert.Equal(t, filepath.Join(appLocation, "var", "lib", "instance-001"),
 				inst.WalDir)
-			assert.Equal(t, filepath.Join(cfgDir, "var", "lib", clusterAppName, "instance-001"),
+			assert.Equal(t, filepath.Join(appLocation, "var", "lib", "instance-001"),
 				inst.VinylDir)
-			assert.Equal(t, filepath.Join(cfgDir, "var", "lib", clusterAppName, "instance-001"),
+			assert.Equal(t, filepath.Join(appLocation, "var", "lib", "instance-001"),
 				inst.MemtxDir)
-			assert.Equal(t, filepath.Join(cfgDir, "var", "run", clusterAppName, "instance-001"),
+			assert.Equal(t, filepath.Join(appLocation, "var", "run", "instance-001"),
 				inst.RunDir)
-			assert.Equal(t, filepath.Join(cfgDir, "var", "run", clusterAppName, "instance-001",
-				"instance-001.control"),
-				inst.ConsoleSocket)
-			assert.Equal(t, "testdata/instances_enabled/cluster_app/config.yml",
-				inst.ClusterConfigPath)
+			assert.Equal(t, filepath.Join(appLocation, "var", "run", "instance-001", "tt.pid"),
+				inst.PIDFile)
+			assert.Equal(t, filepath.Join(appLocation, "var", "run", "instance-001",
+				"tarantool.control"), inst.ConsoleSocket)
+			assert.Equal(t, filepath.Join(appLocation, "var", "log", "instance-001", "tt.log"),
+				inst.Log)
+			assert.Equal(t, filepath.Join(appLocation, "config.yml"), inst.ClusterConfigPath)
 			comparisonsCount++
 
 		case "instance-002":
-			assert.Contains(t, inst.WalDir,
-				"testdata/instances_enabled/cluster_app/instance-002_wal_dir")
-			assert.Contains(t, inst.ConsoleSocket,
-				"testdata/instances_enabled/cluster_app/instance-002.control")
-			assert.Equal(t, filepath.Join(cfgDir, "var", "lib", clusterAppName, "instance-002"),
+			assert.Contains(t, inst.WalDir, filepath.Join(appLocation, "instance-002_wal_dir"))
+			assert.Contains(t, inst.ConsoleSocket, filepath.Join(appLocation,
+				"instance-002.control"))
+			assert.Equal(t, filepath.Join(appLocation, "var", "lib", "instance-002"),
 				inst.VinylDir)
-			assert.Equal(t, filepath.Join(cfgDir, "var", "lib", clusterAppName, "instance-002"),
+			assert.Equal(t, filepath.Join(appLocation, "var", "lib", "instance-002"),
 				inst.MemtxDir)
-			assert.Equal(t, filepath.Join(cfgDir, "var", "run", clusterAppName, "instance-002"),
+			assert.Equal(t, filepath.Join(appLocation, "var", "run", "instance-002"),
 				inst.RunDir)
+			assert.Equal(t, filepath.Join(appLocation, "var", "run", "instance-002", "tt.pid"),
+				inst.PIDFile)
+			assert.Equal(t, filepath.Join(appLocation, "instance-002.control"), inst.ConsoleSocket)
 			comparisonsCount++
 
 		case "instance-003":
-			assert.Contains(t, inst.MemtxDir,
-				"testdata/instances_enabled/cluster_app/instance-003_snap_dir")
-			assert.Contains(t, inst.VinylDir,
-				"testdata/instances_enabled/cluster_app/instance-003_vinyl_dir")
-			assert.Equal(t, filepath.Join(cfgDir, "var", "lib", clusterAppName, "instance-003"),
+			assert.Contains(t, inst.MemtxDir, filepath.Join(appLocation, "instance-003_snap_dir"))
+			assert.Contains(t, inst.VinylDir, filepath.Join(appLocation, "instance-003_vinyl_dir"))
+			assert.Equal(t, filepath.Join(appLocation, "var", "lib", "instance-003"),
 				inst.WalDir)
-			assert.Equal(t, filepath.Join(cfgDir, "var", "run", clusterAppName, "instance-003"),
+			assert.Equal(t, filepath.Join(appLocation, "var", "run", "instance-003"),
 				inst.RunDir)
-			assert.Equal(t, filepath.Join(cfgDir, "var", "run", clusterAppName, "instance-003",
-				"instance-003.control"),
-				inst.ConsoleSocket)
+			assert.Equal(t, filepath.Join(appLocation, "var", "run", "instance-003", "tt.pid"),
+				inst.PIDFile)
+			assert.Equal(t, filepath.Join(appLocation, "var", "run", "instance-003",
+				"tarantool.control"), inst.ConsoleSocket)
 			comparisonsCount++
 
 		default:
@@ -260,4 +265,69 @@ echo "Tarantool 2.11.0"`), 0644)
 	})
 	assert.False(t, canStart)
 	assert.Contains(t, reason, "permission denied")
+}
+
+func Test_collectInstancesForSingleInstApp(t *testing.T) {
+	appName := "script"
+	instancesEnabled, err := filepath.Abs("./testdata/instances_enabled")
+	require.NoError(t, err)
+	appLocation := filepath.Join(instancesEnabled, appName+".lua")
+	apps := []util.AppListEntry{
+		{
+			Name:     appName + ".lua",
+			Location: appLocation,
+		},
+	}
+	appDir := filepath.Join(instancesEnabled, appName)
+	cliOpts := configure.GetDefaultCliOpts()
+	cliOpts.Env.InstancesEnabled = instancesEnabled
+	instances, err := collectInstancesForApps(apps, cliOpts, "/etc/tarantool/")
+	require.NoError(t, err)
+	assert.Equal(t, 1, len(instances))
+
+	inst := instances[0]
+	assert.Equal(t, filepath.Join(appDir, "var", "lib", appName), inst.WalDir)
+	assert.Equal(t, filepath.Join(appDir, "var", "lib", appName), inst.VinylDir)
+	assert.Equal(t, filepath.Join(appDir, "var", "lib", appName), inst.MemtxDir)
+	assert.Equal(t, filepath.Join(appDir, "var", "run", appName), inst.RunDir)
+	assert.Equal(t, filepath.Join(appDir, "var", "run", appName, "tt.pid"), inst.PIDFile)
+	assert.Equal(t, filepath.Join(appDir, "var", "run", appName, "tarantool.control"),
+		inst.ConsoleSocket)
+	assert.Equal(t, filepath.Join(appDir, "var", "log", appName), inst.LogDir)
+	assert.Equal(t, filepath.Join(appDir, "var", "log", appName, "tt.log"), inst.Log)
+	assert.Equal(t, "", inst.ClusterConfigPath)
+	assert.Equal(t, filepath.Join(instancesEnabled, appName), inst.AppDir)
+}
+
+func Test_collectInstancesSingleInstanceTntCtlLayout(t *testing.T) {
+	appName := "script"
+	instancesEnabled, err := filepath.Abs("./testdata/instances_enabled")
+	require.NoError(t, err)
+	appLocation := filepath.Join(instancesEnabled, appName+".lua")
+	apps := []util.AppListEntry{
+		{
+			Name:     appName + ".lua",
+			Location: appLocation,
+		},
+	}
+	cliOpts := configure.GetDefaultCliOpts()
+	cliOpts.Env.InstancesEnabled = instancesEnabled
+	cliOpts.Env.TarantoolctlLayout = true
+	cfgDir := "/etc/tarantool/"
+	instances, err := collectInstancesForApps(apps, cliOpts, cfgDir)
+	require.NoError(t, err)
+	assert.Equal(t, 1, len(instances))
+
+	inst := instances[0]
+	assert.Equal(t, filepath.Join(cfgDir, "var", "lib", appName), inst.WalDir)
+	assert.Equal(t, filepath.Join(cfgDir, "var", "lib", appName), inst.VinylDir)
+	assert.Equal(t, filepath.Join(cfgDir, "var", "lib", appName), inst.MemtxDir)
+	assert.Equal(t, filepath.Join(cfgDir, "var", "run"), inst.RunDir)
+	assert.Equal(t, filepath.Join(cfgDir, "var", "run", appName+".pid"), inst.PIDFile)
+	assert.Equal(t, filepath.Join(cfgDir, "var", "run", appName+".control"),
+		inst.ConsoleSocket)
+	assert.Equal(t, filepath.Join(cfgDir, "var", "log"), inst.LogDir)
+	assert.Equal(t, filepath.Join(cfgDir, "var", "log", appName+".log"), inst.Log)
+	assert.Equal(t, "", inst.ClusterConfigPath)
+	assert.Equal(t, filepath.Join(instancesEnabled, appName), inst.AppDir)
 }